Struct vibrant::Palette

pub struct Palette {
    pub palette: Vec<Rgb<u8>>,
    pub pixel_counts: BTreeMap<usize, usize>,
}

Palette of colors.

Fields

palette

Palette of Colors represented in RGB

pixel_counts

A map of indices in the palette to a count of pixels in approximately that color in the original image.

Methods

impl Palette

fn new<P, G>(image: &G, color_count: usize, quality: i32) -> Palette where P: Sized + Pixel<Subpixel=u8>, G: Sized + GenericImage<Pixel=P>

Create a new palett from an image

Color count and quality are given straight to color_quant, values should be between 8...512 and 1...30 respectively. (By the way: 10 is a good default quality.)

fn sort_by_frequency(&self) -> Self

Change ordering of colors in palette to be of frequency using the pixel count.